AddTolist, Edit

+++ WPF-AddressBook2/MVVM/Views/ContactsView.xaml.cs
@@ -1,25 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 using WPF_AddressBook2.MVVM.Models;
+using WPF_AddressBook2.MVVM.ViewModels;
 using WPF_AddressBook2.Services;
 
 namespace WPF_AddressBook2.MVVM.Views
 {
-    /// <summary>
-    /// Interaction logic for ContactsView.xaml
-    /// </summary>
     public partial class ContactsView : UserControl
     {
         public ContactsView()
@@ -33,13 +19,17 @@ namespace WPF_AddressBook2.MVVM.Views
             var contact = (ContactModel)button.DataContext;
         }
 
+
         private void btn_Remove_Click(object sender, RoutedEventArgs e)
         {
-            var button = (Button)sender;
-            var contact = (ContactModel)button.DataContext;
-
-            ContactService.RemoveFromList(contact);
+            var result = MessageBox.Show("Are you sure you want to delete this contact?", "Confirm Deletion", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
+            {
+                var contact = (sender as Button).DataContext as ContactModel;
+                ContactService.RemoveFromList(contact);
+            }
         }
+
 
         private void ListView_Selected(object sender, RoutedEventArgs e)
         {
